# swagger-auth.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/usuarios": {
            "post": {
                "operationId": "UsuariosController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUsuarioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                },
                "summary": "Crear un nuevo usuario con roles asignados",
                "tags": [
                    "usuarios"
                ]
            },
            "get": {
                "operationId": "UsuariosController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UsuarioResponseDto"
                                    }
                                }
                            }
                        }
                    }
                },
                "summary": "Listar todos los usuarios con sus roles",
                "tags": [
                    "usuarios"
                ]
            }
        },
        "/usuarios/{id}": {
            "get": {
                "operationId": "UsuariosController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "UUID del usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "summary": "Obtener un usuario por ID con sus roles",
                "tags": [
                    "usuarios"
                ]
            },
            "patch": {
                "operationId": "UsuariosController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "UUID del usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUsuarioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "summary": "Actualizar un usuario y sus roles",
                "tags": [
                    "usuarios"
                ]
            },
            "delete": {
                "operationId": "UsuariosController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "UUID del usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario eliminado"
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "summary": "Eliminar un usuario",
                "tags": [
                    "usuarios"
                ]
            }
        },
        "/usuarios/email/{email}": {
            "get": {
                "operationId": "UsuariosController_findByEmail",
                "parameters": [
                    {
                        "name": "email",
                        "required": true,
                        "in": "path",
                        "description": "Email del usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponseDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    }
                },
                "summary": "Buscar usuario por email",
                "tags": [
                    "usuarios"
                ]
            }
        },
        "/usuarios/{id}/roles": {
            "post": {
                "operationId": "UsuariosController_assignRoles",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "UUID del usuario",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rolesIds": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "example": [
                                            "550e8400-e29b-41d4-a716-446655440000"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles asignados exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsuarioResponseDto"
                                }
                            }
                        }
                    }
                },
                "summary": "Asignar roles a un usuario (relación many-to-many)",
                "tags": [
                    "usuarios"
                ]
            }
        },
...
        "/alumnos/usuario/{usuarioId}": {
            "get": {
                "operationId": "AlumnosController_findByUsuario",
                "parameters": [
                    {
                        "name": "usuarioId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alumno encontrado"
                    },
                    "404": {
                        "description": "Alumno no encontrado"
                    }
                },
                "summary": "Obtener alumno por usuario ID",
                "tags": [
                    "Alumnos"
                ]
            }
        },
...
    },
    "info": {
        "title": "PPP Auth Service API",
        "description": "Autenticación y Autorización - Usuarios, Roles, Privilegios",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "schemas": {
            "CreateUsuarioDto": {
                "type": "object",
                "properties": {
                    "nombres": {
                        "type": "string",
                        "description": "Nombres del usuario",
                        "example": "Juan Carlos",
                        "minLength": 2,
                        "maxLength": 120
                    },
                    "apellidos": {
                        "type": "string",
                        "description": "Apellidos del usuario",
                        "example": "Pérez García",
                        "minLength": 2,
                        "maxLength": 120
                    },
                    "email": {
                        "type": "string",
                        "description": "Email del usuario",
                        "example": "juan.perez@upeu.edu.pe",
                        "maxLength": 150
                    },
                    "telefono": {
                        "type": "string",
                        "description": "Teléfono del usuario",
                        "example": "987654321",
                        "maxLength": 20
                    },
                    "contraseña": {
                        "type": "string",
                        "description": "Contraseña del usuario",
                        "example": "Secure@Pass123",
                        "minLength": 6,
                        "maxLength": 255
                    },
                    "activo": {
                        "type": "boolean",
                        "description": "Estado activo del usuario",
                        "example": true,
                        "default": true
                    },
                    "rolesIds": {
                        "description": "IDs de los roles a asignar al usuario (relación many-to-many)",
                        "example": [
                            "550e8400-e29b-41d4-a716-446655440000"
                        ],
                        "items": {
                            "type": "array"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "nombres",
                    "apellidos",
                    "email",
                    "contraseña"
                ]
            },
            "UsuarioResponseDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "ID del usuario",
                        "example": "660e8400-e29b-41d4-a716-446655440000"
                    },
                    "nombres": {
                        "type": "string",
                        "description": "Nombres del usuario",
                        "example": "Juan Carlos"
                    },
                    "apellidos": {
                        "type": "string",
                        "description": "Apellidos del usuario",
                        "example": "Pérez García"
                    },
                    "email": {
                        "type": "string",
                        "description": "Email del usuario",
                        "example": "juan.perez@upeu.edu.pe"
                    },
                    "telefono": {
                        "type": "string",
                        "description": "Teléfono del usuario",
                        "example": "987654321"
                    },
                    "activo": {
                        "type": "boolean",
                        "description": "Estado activo del usuario",
                        "example": true
                    },
                    "roles": {
                        "description": "Roles asignados al usuario (relación many-to-many)",
                        "items": {
                            "type": "array"
                        },
                        "type": "array"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string",
                        "description": "Fecha de creación",
                        "example": "2025-11-18T10:00:00.000Z"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string",
                        "description": "Fecha de actualización",
                        "example": "2025-11-18T10:00:00.000Z"
                    }
                },
                "required": [
                    "id",
                    "nombres",
                    "apellidos",
                    "email",
                    "activo",
                    "roles",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "UpdateUsuarioDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}